requirements.txt
"flask==3.1.0
pandas==2.2.3
python-dotenv==3.12.0" > requirements.txt